+++ Classes/ContentRepository/CreationHandler/DiagramCreationHandler.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Sandstorm\MxGraph;
+namespace Sandstorm\MxGraph\ContentRepository\CreationHandler;
 
 use Neos\ContentRepository\Core\Feature\NodeModification\Dto\PropertyValuesToWrite;
 use Neos\Neos\Ui\Domain\NodeCreation\NodeCreationCommands;
 use Neos\Neos\Ui\Domain\NodeCreation\NodeCreationElements;
 use Neos\Neos\Ui\Domain\NodeCreation\NodeCreationHandlerInterface;
+use Sandstorm\MxGraph\MxGraphConstants;
 
 class DiagramCreationHandler implements NodeCreationHandlerInterface
 {
