+++ Classes/Package.php
@@ -1,19 +1,8 @@
 <?php
 namespace Sandstorm\MxGraph;
 
-use Neos\ContentRepository\Core\Projection\ContentGraph\Node;
-use Neos\Flow\Core\Bootstrap;
 use Neos\Flow\Package\Package as BasePackage;
-
-// TODO: Is this always needed?
 
 class Package extends BasePackage
 {
-    /**
-     * @param Bootstrap $bootstrap The current bootstrap
-     * @return void
-     */
-    public function boot(Bootstrap $bootstrap)
-    {
-    }
 }
